pyproject.toml
[project]
name = "michaelmas"
version = "0.1.0"
description = "A sophisticated, terminal-based AI assistant orchestrated to optimize the digital life, workflows, and information streams of the modern IT professional."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-ai-generativelanguage>=0.9.0",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "langchain>=1.0.8",
    "langchain-google-community>=3.0.1",
    "langchain-google-genai>=3.1.0",
    "langchain-tavily>=0.2.13",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
    "textual>=6.6.0",
]
